Implement mailing list policies and owners.
A mailing list can now be owned by a committee. This committee is then allowed to add or remove subscribers as well as change the name and description of the list.

Each list has a policy determining who can send mail to the list. Options are: everyone, only people subscribed to the list, only *@svcover.nl addresses and only the owning committee email address.