Do the locking manually as this has more or less the intended result
It would have been nicer if I could lock the select, force it to wait but in this implementation at least the last transaction won't bork, but instead will win and overwrite the results of the previous transaction. This can result in multiple fetches of the same facebook data at the moment, but always the latest is stored.